[ISSUE #7784]♻️Extract SendMessage builder helpers
